requirements.txt
# requirements.txt
discord.py>=2.0.0
httpx>=0.25.0
openai>=1.0.0
PyYAML>=6.0
aiosqlite>=0.17.0  # <-- Add for memory DB
google-genai
setuptools